requirements.txt
dash==2.7.0
dash_bootstrap_components==1.4.0
geopy==2.3.0
matplotlib==3.7.2
numpy==1.24.3
pandas==1.5.3
Pillow==10.1.0
plotly==5.9.0
pyproj==3.4.1
requests==2.31.0
seaborn==0.11.2
gunicorn
dash-tools
